Moved dtopotools to src/python directory and updated calls from sloping_fault
