+++ 2d/sloping_fault/setplot_surface.py
@@ -11,7 +11,7 @@ import numpy as np
 from mapc2p import mapc2p
 from plot_okada import plot_okada_surface
 from clawpack.clawutil.data import ClawData
-import dtopotools_horiz_okada as dtopotools
+import clawpack.seismic.dtopotools_horiz_okada as dtopotools
 reload(dtopotools)
 cscale = 8 # scale color limits
